+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Version 2.1.0 (YYYY-MM-DD)
 
 - Add support for vim option `fileencoding` with values `latin1`, `utf-8`, `utf-16` and `utf-16le`.
+- Add support for vim option `fileformat` to force the line ending to `unix`, `dos` or `mac`.
 
 
 ## Version 2.0.0 (2022-11-21)